feat: remove hard-coded assets base asset ids now are dynamically defined according to the network leaving static assets only for mocks, stories and tests
