+++ packages/app/package.json
@@ -25,7 +25,6 @@
     "@fuel-ui/react": "0.23.3",
     "@fuel-ui/test-utils": "0.17.0",
     "@fuel-wallet/connections": "workspace:*",
-    "@fuels/assets": "0.19.0",
     "@fuels/local-storage": "0.19.0",
     "@fuels/react": "0.19.0",
     "@fuels/react-xstore": "0.19.0",
