+++ packages/app/src/systems/Transaction/__mocks__/operation.tsx
@@ -1,11 +1,27 @@
-import type { Operation } from 'fuels';
-import { OperationName, bn } from 'fuels';
+import {
+  type AssetFuel,
+  type Operation,
+  OperationName,
+  assets,
+  bn,
+} from 'fuels';
 
 import { MOCK_TX_RECIPIENT } from './tx-recipient';
 
 // BaseAssetId replacement
 const MOCK_BASE_ASSET_ID =
   '0x0000000000000000000000000000000000000000000000000000000000000000';
+
+export const MOCK_FUEL_ASSETS = assets.map((asset) => {
+  const fuelNetworkAsset = asset.networks.find(
+    (n) => n.type === 'fuel'
+  ) as AssetFuel;
+  return {
+    ...asset,
+    assetId: fuelNetworkAsset.assetId,
+    decimals: fuelNetworkAsset.decimals,
+  };
+});
 
 export const MOCK_OPERATION_CONTRACT_CALL: Operation = {
   name: OperationName.contractCall,
