+++ packages/app/src/systems/Transaction/components/TxOperations/TxOperations.stories.tsx
@@ -1,8 +1,8 @@
 import { Box } from '@fuel-ui/react';
 import { TransactionStatus } from 'fuels';
-import { fuelAssets } from '~/systems/Core';
 
 import {
+  MOCK_FUEL_ASSETS,
   MOCK_OPERATION_CONTRACT_CALL,
   MOCK_OPERATION_TRANSFER,
 } from '../../__mocks__/operation';
@@ -23,7 +23,7 @@ export const Default = (args: TxOperationsProps) => (
       {...args}
       operations={[MOCK_OPERATION_CONTRACT_CALL, MOCK_OPERATION_TRANSFER]}
       status={TransactionStatus.success}
-      assets={fuelAssets}
+      assets={MOCK_FUEL_ASSETS}
     />
   </Box>
 );
